+++ src/EPR.ProducerContentValidation.Application/Constants/FeatureFlags.cs
@@ -3,8 +3,7 @@
     public static class FeatureFlags
     {
         public const string EnableSubsidiaryValidationPom = "EnableSubsidiaryValidationPom";
-        public const string EnableSmallProducerPackagingTypeEnhancedValidation = "EnableSmallProducerPackagingTypeEnhancedValidation";
-        public const string EnableLargeProducerRecyclabilityRatingValidation = "EnableLargeProducerRecyclabilityRatingValidation";
+public const string EnableLargeProducerRecyclabilityRatingValidation = "EnableLargeProducerRecyclabilityRatingValidation";
         public const string EnableLargeProducerEnhancedRecyclabilityRatingValidation = "EnableLargeProducerEnhancedRecyclabilityRatingValidation";
     }
 }
